+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.0] - 2022-05-28
+
+## Changed
+
+- Dropped support for Python 3.6 & Django 3.1 since those are now longer maintained
+- Added support for Django 4.0
+
 ## [1.1.0] - 2021-11-02
 
 ## Added